Entrypoint (#818)
* changes

* address todos

* feat: MinIO for worlds
feat: add minio-compose.yml for helpful MinIO server

* chore: delete RuntimeDependencyManager since plugins are gone

* chore: rename InstanceManager to WorldManager to better reflect its task

* chore: clean up build.gradle.kts

* chore: cleanup CytosisBootstrap and remove benchmark manager

* chore: even more cleanup

* feat: use new redis api (old one was deprecated)

* chore: fix some things qodana pointed out

* feat: launcher module to just run cytosis

* chore: use jetbrains annotations instead of jspecify

* feat: add proxy support
chore: update to latest entrypoint

* fix: update database schema

* fix: stop command not working

* feat: move world api from entrypoint to cytosis

* chore: update to latest entrypoint

* bump version

* move MetricsConfig to CytosisConfig

* chore: use lombok getters

* chore: use Cytosis.getGeneric for sideboard and player list

* chore: add @Blocking

* chore: update to latest entrypoint

* feat: change to garage from minio

* fix: notify listeners not being registered

* fix: prevent yoink command from yoinking players already on your server

* fix: get the player's instance for time command

* fix: snooper help messages missing spaces

* fix: require instance for npc

* fix: polar world importer outdated help message

* fix: update ranks for nicks

* feat: InstanceResolver to fetch premade instances of classes

* fix: all chat not being ignored

* fix: more safe guards for find command

* chore: remove uneeded params in commands

* feat: ToggleablePreference for all of your toggleable preferences neededs

* fix: use Msg.red

* fix: register instance on create

+++ src/main/java/net/cytonic/cytosis/managers/SideboardManager.java
@@ -6,7 +6,6 @@ import java.util.UUID;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import net.minestom.server.MinecraftServer;
 import net.minestom.server.timer.Task;
 import net.minestom.server.timer.TaskSchedule;
@@ -18,7 +17,8 @@ import net.cytonic.cytosis.bootstrap.annotations.CytosisComponent;
 import net.cytonic.cytosis.data.objects.ExpiringMap;
 import net.cytonic.cytosis.nicknames.NicknameManager;
 import net.cytonic.cytosis.player.CytosisPlayer;
-import net.cytonic.cytosis.sideboard.DefaultCreator;
+import net.cytonic.cytosis.server.AbstractCytosisServer;
+import net.cytonic.cytosis.server.sideboard.SideboardService;
 import net.cytonic.cytosis.sideboard.Sideboard;
 import net.cytonic.cytosis.sideboard.SideboardCreator;
 
@@ -27,19 +27,23 @@ import net.cytonic.cytosis.sideboard.SideboardCreator;
  */
 @NoArgsConstructor
 @CytosisComponent(dependsOn = {NicknameManager.class})
-public class SideboardManager implements Bootstrappable {
+public class SideboardManager<P extends CytosisPlayer> implements Bootstrappable {
 
     private final Map<UUID, Sideboard> sideboards = new ExpiringMap<>();
     @Getter
     @Nullable
     private Task task = null;
-    @Getter
-    @Setter
-    private SideboardCreator sideboardCreator = new DefaultCreator();
+    private SideboardCreator<P> sideboardCreator;
+    private SideboardService<P> sideboardService;
 
     @Override
     public void init() {
-        autoUpdateBoards(TaskSchedule.seconds(1));
+        sideboardService = Cytosis.<AbstractCytosisServer<P>>getGeneric(AbstractCytosisServer.class)
+            .sideboardService();
+        if (sideboardService.supportsSideboard()) {
+            sideboardCreator = sideboardService.sideboardCreator();
+            autoUpdateBoards(sideboardService.scheduale());
+        }
     }
 
     @Override
@@ -52,7 +56,8 @@ public class SideboardManager implements Bootstrappable {
      *
      * @param player the player
      */
-    public void addPlayer(CytosisPlayer player) {
+    public void addPlayer(P player) {
+        if (!sideboardService.supportsSideboard()) return;
         sideboards.put(player.getUuid(), sideboardCreator.sideboard(player));
     }
 
@@ -64,8 +69,9 @@ public class SideboardManager implements Bootstrappable {
     }
 
     public void updatePlayersNow() {
+        if (!sideboardService.supportsSideboard()) return;
         sideboards.forEach((uuid, sideboard) -> {
-            Optional<CytosisPlayer> player = Cytosis.getPlayer(uuid);
+            Optional<P> player = Cytosis.getPlayer(uuid);
             if (player.isEmpty()) {
                 sideboard.delete();
                 sideboards.remove(uuid);
